Move ArrowTooltip to separate file (#51)
